requirements-dev.txt
# Include base requirements
-r requirements.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.2.0

# Code Quality
black==24.10.0
flake8==7.1.1
isort==5.13.2
pylint==3.3.4

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.2
